Restoring C
