+++ threestudio/utils/misc.py
@@ -6,7 +6,7 @@ import tinycudann as tcnn
 import torch
 from packaging import version
 
-from threestudio.utils.config import C, config_to_primitive
+from threestudio.utils.config import config_to_primitive
 from threestudio.utils.typing import *
 
 
@@ -60,6 +60,30 @@ def load_module_weights(
             state_dict_to_load[m.group(1)] = v
 
     return state_dict_to_load, ckpt["epoch"], ckpt["global_step"]
+
+
+def C(value: Any, epoch: int, global_step: int) -> float:
+    if isinstance(value, int) or isinstance(value, float):
+        pass
+    else:
+        value = config_to_primitive(value)
+        if not isinstance(value, list):
+            raise TypeError("Scalar specification only supports list, got", type(value))
+        if len(value) == 3:
+            value = [0] + value
+        assert len(value) == 4
+        start_step, start_value, end_value, end_step = value
+        if isinstance(end_step, int):
+            current_step = global_step
+            value = start_value + (end_value - start_value) * max(
+                min(1.0, (current_step - start_step) / (end_step - start_step)), 0.0
+            )
+        elif isinstance(end_step, float):
+            current_step = epoch
+            value = start_value + (end_value - start_value) * max(
+                min(1.0, (current_step - start_step) / (end_step - start_step)), 0.0
+            )
+    return value
 
 
 def cleanup():
